requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
boto3==1.43.7
botocore==1.43.7
celery==5.3.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cohere==5.5.8
colorama==0.4.6
coverage==7.5.4
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.30
django-cors-headers==4.9.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
factory_boy==3.3.3
Faker==40.15.0
fastavro==1.12.2
filelock==3.29.0
fsspec==2026.4.0
groq==1.2.0
h11==0.16.0
hf-xet==1.5.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.14.0
idna==3.13
inflection==0.5.1
iniconfig==2.3.0
jiter==0.14.0
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
markdown-it-py==4.2.0
mdurl==0.1.2
openai==1.30.5
packaging==26.2
parameterized==0.9.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
redis==5.0.8
referencing==0.37.0
requests==2.34.1
rich==15.0.0
rpds-py==0.30.0
s3transfer==0.17.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tokenizers==0.23.1
tqdm==4.67.3
typer==0.25.1
types-requests==2.33.0.20260513
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
uritemplate==4.2.0
urllib3==2.7.0
vine==5.1.0
wcwidth==0.7.0
